refactor(platform-wallet): rename PlatformWalletManager to PlatformWalletCoordinator
Avoids confusion with key-wallet's WalletManager which manages keys,
accounts, and UTXOs. PlatformWalletCoordinator coordinates the SDK,
wallet manager, SPV, and event handling — it's the orchestration layer.
